Reworking the APIClient (#79)
* Removing api definitions

* Testing extending APIClient

* Exporting as BaseAPIClient now

* Fixing tests

* Update package.json
