+++ src/api/client.ts
@@ -1,17 +1,13 @@
 import {APIProvider, APIResponse, FetchProvider, FetchProviderOptions} from './provider'
 import {ABISerializableConstructor, ABISerializableType} from '../serializer/serializable'
 import {abiDecode} from '../serializer/decoder'
-import {ChainAPI} from './v1/chain'
-import {HistoryAPI} from './v1/history'
 import {BuiltinTypes} from '../serializer/builtins'
 
-export {ChainAPI, HistoryAPI}
-
 export interface APIClientOptions extends FetchProviderOptions {
-    /** URL to the API node to use, only used if the provider option is not set. */
-    url?: string
     /** API provider to use, if omitted and the url option is set the default provider will be used.  */
     provider?: APIProvider
+    /** URL to the API node to use, only used if the provider option is not set. */
+    url?: string
 }
 
 export interface APIErrorDetail {
@@ -92,7 +88,7 @@ export class APIError extends Error {
     }
 }
 
-export class APIClient {
+export class BaseAPIClient {
     static __className = 'APIClient'
 
     readonly provider: APIProvider
@@ -105,11 +101,6 @@ export class APIClient {
         } else {
             throw new Error('Missing url or provider')
         }
-    }
-
-    v1 = {
-        chain: new ChainAPI(this),
-        history: new HistoryAPI(this),
     }
 
     async call<T extends ABISerializableConstructor>(args: {
