--- src/api/types.ts
@@ -1,1 +1,0 @@
-export * as v1 from './v1/types'
